Modifications to the datatypes defined in the migrations file to standardise

+++ database/migrations/2020_02_02_123931_create_transactions_table.php
@@ -15,14 +15,18 @@ class CreateTransactionsTable extends Migration
     {
         Schema::create('transactions', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('quantity')->unsigned();
-            $table->integer('buyer_id')->unsigned();
-            $table->integer('product_id')->unsigned();
+            $table->bigInteger('quantity')->unsigned();
+            $table->bigInteger('buyer_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
             $table->timestamps();
 
             $table->foreign('buyer_id')->references('id')->on('users');
             $table->foreign('product_id')->references('id')->on('products');
         });
+
+        // Schema::create('transactions', function (Blueprint $table) {
+        //
+        // });
     }
 
     /**
