+++ database/migrations/2020_02_02_152635_category_product_table.php
@@ -14,8 +14,8 @@ class CategoryProductTable extends Migration
     public function up()
     {
         Schema::create('category_product', function (Blueprint $table) {
-            $table->integer('category_id')->unsigned();
-            $table->integer('product_id')->unsigned();
+            $table->bigInteger('category_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
 
             // Define the foreign keys
             $table->foreign('category_id')->references('id')->on('categories');
